Orchestration_Frameworks/LangGraph/pyproject.toml
[project]
name = "langgraph-local-agent"
version = "0.1.0"
description = "A simple LangGraph service using a local Llama agent"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "langgraph",
    "langchain",
    "langchain-community",
    "langchain-openai",
    "python-dotenv",
]

[tool.uv]
# No specific uv settings needed for now, but section can exist
